requirements.txt
certifi>=2020.12.5
idna>=2.10
jsons>=1.6.3
requests>=2.25.1
pycryptodome>=3.12.0
typish==1.9.3
urllib3>=1.26.5
aiohttp>=3.7.4.post0
chardet==5.0.0